Bump up instance count
Docs claim to expect "up to 30s" load times, but ours isn't responding
at all. It says it's because redis, but if that's the case why did it
work just fine before swarming it with the load balancing stuff?

Convinced changing tag from 'precompose' to 'latest' is messing with it.